// frontend/src/types/UserInfo.ts
export default interface UserInfo {
  name: string
  userid: number
  backgd_url: string
  avtr_url: string
  email: string
  phone: string
  signiture: string
}

// accountNonExpired: true
// accountNonLocked: true
// authorities: [{authority: "ROLE_USER"}]
// avatar: "上传图片大小必须在4MB以内！"
// backImg: "https://images.maiquer.tech/images/bef07109.jpeg"
// createTime: "2022-04-11 14:26:14"
// credentialsNonExpired: true
// email: null
// enable: true
// enabled: true
// giftEvaluations: []
// id: 99
// likeEvaluations: [,…]
// myEvaluations: [,…]
// nickname: "Dustella"
// openid: "o0F3X0w8bAJsMnTWTxPuP441X2eE"
// password: null
// phone: "13770808113"
// powerList: null
// roleSet: [,…]
// signature: "从前从前，有个人爱你很久"
// status: "normal"
// unionid: "okXCn5jykNPHKjbusklowmvokeMY"
// updateTime: "2022-04-22 11:24:07"
// username: "d3cb6c87be"